Move resolver tests to separate files (one per resolver class).
